// apps/server/src/types.ts
export type Env = {
	DB: D1Database;
	GITHUB_API_KEY: string;
	KV: KVNamespace;
};

export interface Author {
	login: string;
	id: number;
	node_id: string;
	avatar_url: string;
	gravatar_id: string;
	url: string;
	html_url: string;
	followers_url: string;
	following_url: string;
	gists_url: string;
	starred_url: string;
	subscriptions_url: string;
	organizations_url: string;
	repos_url: string;
	events_url: string;
	received_events_url: string;
	type: string;
	site_admin: boolean;
}

export interface Release {
	url: string;
	assets_url: string;
	upload_url: string;
	html_url: string;
	id: number;
	author_id: string;
	node_id: string;
	repo_id: string;
	tag_name: string;
	target_commitish: string;
	name: string;
	draft: boolean;
	prerelease: boolean;
	created_at: string;
	published_at: string;
	assets: unknown[];
	tarball_url: string;
	zipball_url: string;
	body: string;
}

export type GitHubRelease = Omit<Release, "author_id"> & {
	author: Author;
};